Feat: Components/carrossel-oficinas/ Implementando o carrossel @gabrielapsgomes @angeloneves19
